\name{nmRIO-package}
\alias{nmRIO-package}
\alias{nmRIO}
\docType{package}
\title{Parser for NMR raw fid and processed data from nmrML, Bruker and Varian instruments}
\description{
The package reads NMR data into R objects, 
following the data model in nmrML.
Bruker and Varian files are only supported to a limited extent. 
}
\details{
\tabular{ll}{
Package: \tab nmRIO\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-10-09\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
